+++ components/Experiences.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import './Experiences.scss';
+
+export default () => (
+  <div className="header">
+    <h1>helloimela.com</h1>
+  </div>
+);
